+++ ocsp.go
@@ -106,8 +106,6 @@ const (
 	maxClockSkew           = 900 * time.Second // buffer for clock skew
 )
 
-var stopOCSPCacheClearing = make(chan struct{}, 2)
-
 type ocspStatusCode int
 
 type ocspStatus struct {
@@ -1121,8 +1119,9 @@ func initOcspModule() {
 }
 
 type ocspCacheClearerType struct {
-	running bool
 	mu      sync.Mutex
+	running bool
+	cancel  context.CancelFunc
 }
 
 func (occ *ocspCacheClearerType) start() {
@@ -1131,6 +1130,8 @@ func (occ *ocspCacheClearerType) start() {
 	if occ.running {
 		return
 	}
+	ctx, cancel := context.WithCancel(context.Background())
+	occ.cancel = cancel
 	interval := defaultOCSPResponseCacheClearingInterval
 	if intervalFromEnv := os.Getenv(ocspResponseCacheClearingIntervalInSecondsEnv); intervalFromEnv != "" {
 		intervalAsSeconds, err := strconv.Atoi(intervalFromEnv)
@@ -1147,12 +1148,11 @@ func (occ *ocspCacheClearerType) start() {
 			select {
 			case <-ticker.C:
 				clearOCSPCaches()
-			case <-stopOCSPCacheClearing:
+			case <-ctx.Done():
 				occ.mu.Lock()
 				defer occ.mu.Unlock()
 				logger.Debug("stopped clearing OCSP cache")
 				ticker.Stop()
-				stopOCSPCacheClearing <- struct{}{}
 				occ.running = false
 				return
 			}
@@ -1163,11 +1163,9 @@ func (occ *ocspCacheClearerType) start() {
 
 func (occ *ocspCacheClearerType) stop() {
 	occ.mu.Lock()
-	running := occ.running
-	occ.mu.Unlock()
-	if running {
-		stopOCSPCacheClearing <- struct{}{}
-		<-stopOCSPCacheClearing
+	defer occ.mu.Unlock()
+	if occ.running {
+		occ.cancel()
 	}
 }