+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Bug fixes:
 - Added panic recovery block for stage file uploads and downloads operation (snowflakedb/gosnowflake#1687).
 - Fixed WIF metadata request from Azure container, manifested with HTTP 400 error (snowflakedb/gosnowflake#1701).
 - Fixed SAML authentication port validation bypass in `isPrefixEqual` where the second URL's port was never checked (snowflakedb/gosnowflake#1712).
+- Fixed a race condition in OCSP cache clearer (snowflakedb/gosnowflake#1704).
 
 ## 1.19.0